consolidate notification tests
